pyproject.toml
[tool.poetry]
name = "melanoma-detection"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
matplotlib = "^3.9.1"
pillow = "^10.4.0"
display = "^1.0.0"
opendatasets = "^0.1.22"
opencv-python = "^4.10.0.84"
pathlib = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
